IMPALA-12780: Only show non-default options in the catalog operations page
The details shown in the catalog operations page are too verbose. For
instance, for ExecDdlRequest, we show the string of TDdlQueryOptions:
  query_options=TDdlQueryOptions(sync_ddl:false, debug_action:,
lock_max_wait_time_s:300, kudu_table_reserve_seconds:0)
What really matters is the non-default options, e.g. if sync_ddl is set
to true, it should be shown.

This patch improves the details field to show only non-default options.

Also wraps the content of Query Id and Details columns so they can fit
into the table.

Change-Id: Ie62d4b9e4d357e02764e7a62f4dc107de602e1a5
Reviewed-on: http://gerrit.cloudera.org:8080/20985
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>