IMPALA-12781: ARRAY<STRUCT<s: STRING> crashes in top-n
In ASAN builds, and if codegen is enabled, if we sort an array
containing a struct of a string with limit, Impala crashes.

This is because of an error in
SlotDescriptor::CodegenWriteCollectionStructChild(). This function is
responsible for copying the variable length data of struct children, but
currently we return early without doing anything if
'children_tuple_desc->string_slots()' and
'children_tuple_desc->collection_slots()' are empty. However, this is
incorrect because the variable length children of a struct are listed in
the corresponding members of the struct's parent slot, not the struct
slot itself; see the comment in TupleDescriptor::AddSlot(). Because of
this, the variable length data of the StringValue is not copied.

This commit removes the early return from
SlotDescriptor::CodegenWriteCollectionStructChild() and inserts a DCHECK
in TupleDescriptor::string_slots() and
TupleDescriptor::collection_slots() that ensures that these functions
are not called on item tuples of structs, as the result would be
unexpected/incorrect.

This change also adjusts some variable names in
SlotDescriptor::CodegenWriteCollectionToSlot() to make them clearer and
updates the codegen IR example of
SlotDescriptor::CodegenWriteStringOrCollectionToSlot().

Testing:
 - added a query in top-n-complex.test that used to crash, and the same
   query without LIMIT in sort-complex.test for completeness.

Change-Id: If87d9e44775e809da9a13e953b4d2c3db9728801
Reviewed-on: http://gerrit.cloudera.org:8080/20988
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>