IMPALA-12782: Show info of the event processing in /events webUI
The /events page of catalogd shows the metrics and status of the
event-processor. This patch adds more info in this page, including
 - lag info
 - current event batch that's being processed
See the screenshot attached in the JIRA for how it looks like.

Also moves the error message to the top to highlight the error status.
Fixes the issue of not updating latest event id when event processor is
stopped. Also fixes the issue of error message not cleared after global
INVALIDATE METADATA.

Adds a debug action, catalogd_event_processing_delay, to inject a sleep
while processing an event. So the web page can be captured more easily.

Also adds a missing test for showing the error message of
event-processing in the /events page.

Tests:
 - Add e2e test to verify the content of the page.

Change-Id: I2e7d4952c7fd04ae89b6751204499bf9dd99f57c
Reviewed-on: http://gerrit.cloudera.org:8080/20986
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>