IMPALA-12787: Concurrent DELETE and UPDATE operations on Iceberg tables can be problematic
If an UPDATE operation runs concurrently with a DELETE operation, and
the DELETE commits first, then the UPDATE can revive deleted rows. This
is because only RowDelta.validateNoConflictingDataFiles() is called, but
RowDelta.validateNoConflictingDeleteFiles() is not. Therefore, the
UPDATE operation ignores the concurrently written delete files.

This patch adds RowDelta.validateNoConflictingDeleteFiles() to UPDATE
operations.

Testing
 * added a stress test to validate concurrent DELETE and UPDATE
    operations

Change-Id: I9e581ea17fa8f6ccd9c87aaad1281bb694079f6e
Reviewed-on: http://gerrit.cloudera.org:8080/20999
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>