IMPALA-12755: Disable snapshots from cdh.rcs.releases.repo
This is a 3.x-only change.

There is no longer the pom file for cdh-root in version 6.x-SNAPSHOT at
https://repository.cloudera.com/artifactory/cdh-releases-rcs
This patch disables using snapshots from this repo so the pom file can
be downloaded from the native-toolchain S3 bucket.

This also bumps the Maven version installed by bootstrap_system.sh and
bootstrap_build.sh to v3.9.2 since the older version can't be
downloaded.

Test:
 - Run CORE tests

Change-Id: I14b6aa49ee6e877b5020348d7bddb00f88a284b9
Reviewed-on: http://gerrit.cloudera.org:8080/20995
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Quanlong Huang <huangquanlong@gmail.com>