IMPALA-11568: Fix SamplingTimeSeriesCounter not correctly cleared
TimeSeriesCounters are scheduled in PeriodicCounterUpdater at the time
when they are added. TimeSeriesCounterToJsonTest stops and clears them
to avoid the update thread interfere with the rest of the test. However,
the samples in SamplingTimeSeriesCounter are not cleared since it
doesn't implement the Clear() method of TimeSeriesCounter. This leads to
non-determined results depend on when the counter is stopped.

This patch adds a Reset() method for SamplingTimeSeriesCounter to clear
the samples. It's only used in the test. A Reset() method in
StreamingSampler is also added for this purpose.

Test:
 - Ran TimeSeriesCounterToJsonTest 100 times. It used to fail in 10 runs.

Change-Id: I8bc2c94ad96899522b288708fcaf7413d4b99fa1
Reviewed-on: http://gerrit.cloudera.org:8080/21001
Reviewed-by: Riza Suminto <riza.suminto@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>