IMPALA-12795: Deflake test_catalog_operation_fields
test_catalog_operation_fields verifies the CREATE DATABASE statement has
consistent fields in the catalog operations page and the impalad queries
page. It could fail if the query is waiting to be closed (so not in the
list of completed queries). This fix the test by also finding the query
in the list of in-flight queries.

Tests:
 - Ran the test 18000 times with other concurrent tests running

Change-Id: Ibc8e642d560c1a4ca1bdc3a618eb67e1329b47d2
Reviewed-on: http://gerrit.cloudera.org:8080/21023
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>