IMPALA-12808: test_iceberg_deletes_and_updates failed by timeout error
On some systems test_iceberg_deletes_and_updates fails because
the delete thread cannot succeed as it always gets validation
errors due to the concurrent UPDATEs.

This patch adds SYNC_DDL=true to the DELETE/UPDATE threads,
and also prints out the exceptions which can be useful to debug
test failures in the future.

Testing:
 * validated the fix in an environment where the test failed before

Change-Id: Icb0b078df2d95f3a62b124b6148f2f32aa425be4
Reviewed-on: http://gerrit.cloudera.org:8080/21033
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>