IMPALA-12816: Remove exclusion of Netty to allow SSL connections to HBase
It is possible to configure Impala to use SSL when connecting to HBase.
Under the covers, an HBase client connection also has a connection to
ZooKeeper. SSL connections to ZooKeeper require Netty. Netty has been
excluded from Impala’s classpath since IMPALA-9648 because of earlier
security problems. IMPALA-9648 says that we need to be at least at
version 4.1.44 of Netty and we now pull in 4.1.86. To allow use of SSL
with HBase, remove the exclusion of Netty.

TESTING:

Ran end-to-end tests.
Ran queries against HBase configured to use SSL in a Cloudera
deployment.

Change-Id: Ic3038369345ca74f01ce2d4248c5beb21b633b2b
Reviewed-on: http://gerrit.cloudera.org:8080/21034
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>