IMPALA-12811: Exception during re-analyze can be lost
When there is an AnalysisException during re-analyze, we try to print
the re-written and original statements via invoking toSql() on them.
But toSql() fails because the analysis of the statement was incomplete,
so it throws another exception (typically an IllegalStateException
without any relevant information about the original issue).

This patch puts the original LOG.error() in a try-catch, then wraps the
original AnalysisException into a new exception that just mentions that
the error occurred after query rewrite.

Testing:
 * added a column masking test with an invalid masking function

Change-Id: Ie6e36b08703c07a2a8d68a4ec0e8ddd65ba03199
Reviewed-on: http://gerrit.cloudera.org:8080/21037
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>