IMPALA-12823: Fix repeated query not found messages in impalad.INFO logs
If an unknown or closed query id is requested from the impala
server, the page repeatedly queries the server producing repeated
query not found messages in the impalad.INFO logs.

The behaviour is seen in query plan and query summary pages.

This has been fixed by stopping further requests on receiving
an error response from the server.

Change-Id: I6d994482b6f67a2d09c6a924eaaf90b230245b61
Reviewed-on: http://gerrit.cloudera.org:8080/21052
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>