IMPALA-12842: Make node.js installation aware of CPU architecture
The earlier version of the installation logic in
tests/run-js-test-on-arm.sh always used 'linux-x86' as the architecture
tag for the downloaded node.js installation package, which made the
tests themselves fail when running on an instance powered by an ARM
(Graviton) CPU.

This patch uses the $ARCH_NAME environment variable set up in
bin/impala-config.sh to add the correct CPU architecture tag to the URL
of the node.js package to be downloaded. If the rported CPU type is not
one of the known types (x86_64 or aarch64), the installation step throws
and error and emits a JUnit XML symptom for Jenkins runs.

Change-Id: Ie22da6237ed6d19cf8af721df471033fdfe6b23a
Reviewed-on: http://gerrit.cloudera.org:8080/21070
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Surya Hebbar <shebbar@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>