IMPALA-12717: Fix error message for missing table descriptor
The error messages for missing table descriptor were malformed due to
incorrect stringstream usage. The message was given in the constructor
and the table id overwrote the beginning the following way:
'ERROR: 0ailed to get table descriptor for table id: '
The solution is not passing any string to the std::stringstream
constructor and using only operator<< instead.

Testing: Built a version locally where the table descriptor was
missing from Iceberg DELETE statement, and verified that the error
message is 'ERROR: Failed to get table descriptor for table id: 0'.

Change-Id: Iae94efe7f60a2de7f4d9ec47ad9d2c1ee8ad0b41
Reviewed-on: http://gerrit.cloudera.org:8080/21084
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>