IMPALA-12730: Don't use -Weverything for clang-tidy
For current clang-tidy builds, we enable -Weverything and then
explicitly disable a couple verbose warnings. Clang-tidy uses
these warnings to produce its clang-diagnostic-* issues.
This gives clang-tidy the maximum flexibility to enforce any
warning the Clang can produce, but in practice, we disable
them via the .clang-tidy file's Checks: section.

The output from a -Weverything build is enormous and hard to
navigate. When there is a compilation issue, digging through
the logs is arduous. Clang discourages the use of -Weverything,
as it contains many noisy warnings.

This changes the clang-tidy build to use -Wextra instead and
shrinks the list of disabled clang-diagnostic-* issues
accordingly.

Testing:
 - Verified that bin/run_clang_tidy.sh passes

Change-Id: Ib4f4cc9a6d040b90495496b2396a26f3c5189231
Reviewed-on: http://gerrit.cloudera.org:8080/21112
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>