IMPALA-12611: Add support to MAP type Iceberg Metadata table columns
This change adds support for querying MAP types from Iceberg Metadata
tables.

The 'IcebergMetadataScanner.ArrayScanner' java class is renamed to
'CollectionScanner' and extended to be able to handle maps. For arrays
the iteration returns the element as before, for maps it returns
'Map.Entry' objects.

Note that collections in the FROM clause are still not supported.

Testing:
- Added E2E tests in iceberg-metadata-tables.test.

Change-Id: I8a8b3a574ca45c893315c3b41b33ce4e0eff865a
Reviewed-on: http://gerrit.cloudera.org:8080/21125
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>